dashboard/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.3
django-cors-headers==1.2.2
django-filter==0.15.3
django-memoize==2.0.0
django-rest-auth==0.8.2
django-reversion==2.0.8
djangorestframework==3.5.2
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.9.1
djangorestframework-jwt==1.8.0
jsonfield==1.0.3
jsonschema==2.5.1
kombu==4.0.2
PyJWT==1.4.2
PyMySQL==0.7.9
pytz==2016.7
requests==2.11.1
rest-condition==1.0.3
rest-framework-generic-relations==1.1.0
six==1.10.0
vine==1.1.3
redis==2.10.5
